+++ code/Terrain/Editor/ElevateBrush.cpp
@@ -21,7 +21,7 @@ ElevateBrush::ElevateBrush(const resource::Proxy< hf::Heightfield >& heightfield
 {
 }
 
-uint32_t ElevateBrush::begin(int32_t x, int32_t y, const State& state)
+uint32_t ElevateBrush::begin(float x, float y, const State& state)
 {
 	m_radius = state.radius;
 	m_fallOff = state.falloff;
@@ -30,7 +30,7 @@ uint32_t ElevateBrush::begin(int32_t x, int32_t y, const State& state)
 	return MdHeight;
 }
 
-void ElevateBrush::apply(int32_t x, int32_t y)
+void ElevateBrush::apply(float x, float y)
 {
 	for (int32_t iy = -m_radius; iy <= m_radius; ++iy)
 	{
@@ -65,7 +65,7 @@ void ElevateBrush::apply(int32_t x, int32_t y)
 	}
 }
 
-void ElevateBrush::end(int32_t x, int32_t y)
+void ElevateBrush::end(float x, float y)
 {
 }
