+++ code/Terrain/Editor/NoiseBrush.cpp
@@ -25,7 +25,7 @@ NoiseBrush::NoiseBrush(const resource::Proxy< hf::Heightfield >& heightfield)
 {
 }
 
-uint32_t NoiseBrush::begin(int32_t x, int32_t y, const State& state)
+uint32_t NoiseBrush::begin(float x, float y, const State& state)
 {
 	m_radius = state.radius;
 	m_fallOff = state.falloff;
@@ -33,7 +33,7 @@ uint32_t NoiseBrush::begin(int32_t x, int32_t y, const State& state)
 	return MdHeight;
 }
 
-void NoiseBrush::apply(int32_t x, int32_t y)
+void NoiseBrush::apply(float x, float y)
 {
 	for (int32_t iy = -m_radius; iy <= m_radius; ++iy)
 	{
@@ -52,7 +52,7 @@ void NoiseBrush::apply(int32_t x, int32_t y)
 	}
 }
 
-void NoiseBrush::end(int32_t x, int32_t y)
+void NoiseBrush::end(float x, float y)
 {
 }
